+++ framework/CHANGELOG.md
@@ -151,6 +151,7 @@ Yii Framework 2 Change Log
 - Enh #4297: Added check for DOM extension to requirements (samdark)
 - Enh #4317: Added `absoluteAuthTimeout` to yii\web\User (ivokund, nkovacs)
 - Enh #4360: Added client validation support for file validator (Skysplit)
+- Enh #4433: Made Formatter asSize function more robust, added all defined prefixes and added a maximumPrefix parameter (kmindi)
 - Enh: Added support for using sub-queries when building a DB query with `IN` condition (qiangxue)
 - Enh: Supported adding a new response formatter without the need to reconfigure existing formatters (qiangxue)
 - Enh: Added `yii\web\UrlManager::addRules()` to simplify adding new URL rules (qiangxue)